docs: update README for v0.0.5 changes
- Remove .chapter-urls.json references (system removed)
- Add check-new/check-only API endpoints
- Add clear/pending endpoint
- Update chapter tracking to reflect DB + filesystem approach
- Add multi-group support mention
- Add fast parallel checking description

+++ komga/src/main/kotlin/org/gotson/komga/infrastructure/metadata/comicrack/dto/ComicInfo.kt
@@ -19,7 +19,7 @@ class ComicInfo {
   var count: Int? = null
 
   @JsonProperty(value = "Volume")
-  var volume: Int? = null
+  var volume: Float? = null
 
   @JsonProperty(value = "AlternateSeries")
   var alternateSeries: String? = null
